+++ packages/cli/src/utils/hosting/hosting.js
@@ -36,11 +36,10 @@ class HostingFile {
 }
 
 class Hosting {
-  constructor (hostingName, socket) {
-    debug('Hosting.constructor', hostingName, socket)
+  constructor (hostingName) {
+    debug('Hosting.constructor', hostingName)
 
     this.name = hostingName
-    this.socket = socket || null
     this.path = null
 
     this.existRemotely = null
@@ -50,6 +49,11 @@ class Hosting {
     this.editHostingURL = `https://${session.getHost()}${this.hostingURL}${this.name}/`
     this.hostingHost = session.getHost() === 'api.syncano.rocks' ? 'syncano.ninja' : 'syncano.site'
     this.config = {}
+
+    // Remote state
+    this.remote = {
+      domains: []
+    }
 
     this.loadLocal()
   }
@@ -91,27 +95,45 @@ class Hosting {
   }
 
   hasCNAME (cname) {
-    return this.domains.indexOf(cname) > -1
+    return this.remote.domains.indexOf(cname) > -1
+  }
+
+  updateHosting () {
+    const params = {
+      src: this.src,
+      cname: this.cname,
+      config: {
+        browser_router: this.browser_router || false
+      }
+    }
+    if (!this.cname) {
+      delete params.cname
+    }
+    session.settings.project.updateHosting(this.name, params)
   }
 
   async configure (params) {
-    const domains = this.domains
-    if (params.cname && this.domains.indexOf(params.cname) < 0) {
+    const domains = this.remote.domains
+    if (params.cname && domains.indexOf(params.cname) < 0) {
       domains.push(params.cname)
     }
 
     if (params.removeCNAME) {
+      this.cname = null
       const cnameToRemoveIndex = domains.indexOf(params.removeCNAME)
       if (cnameToRemoveIndex > -1) {
         domains.splice(cnameToRemoveIndex, 1)
       }
     }
 
+    this.cname = params.cname
+    this.domains = domains
+    this.config.browser_router = params.browser_router
+    this.updateHosting()
+
     const paramsToUpdate = {
       name: this.name,
-      config: {
-        browser_router: params.browser_router
-      },
+      config: this.config,
       domains
     }
 
@@ -123,7 +145,6 @@ class Hosting {
         'X-Api-Key': session.settings.account.getAuthKey()
       }
     })
-
     return this.setRemoteState(response.data)
   }
 
@@ -142,7 +163,9 @@ class Hosting {
     // TODO: not optimal
     const paramsToUpdate = {
       name: this.name,
-      domains: this.domains
+      domains: this.domains,
+      config: this.config,
+      auth: this.auth || {}
     }
 
     const response = await axios.request({
@@ -158,27 +181,21 @@ class Hosting {
   }
 
   async delete () {
-    if (!this.socket) {
-      await axios.request({
-        url: this.editHostingURL,
-        method: 'DELETE',
-        headers: {
-          'X-Api-Key': session.settings.account.getAuthKey()
-        }
-      })
-      session.settings.project.deleteHosting(this.name)
-      return this
-    }
+    await axios.request({
+      url: this.editHostingURL,
+      method: 'DELETE',
+      headers: {
+        'X-Api-Key': session.settings.account.getAuthKey()
+      }
+    })
+    session.settings.project.deleteHosting(this.name)
+    return this
   }
 
-  static get (hostingName, socket) {
+  static get (hostingName) {
     debug(`get ${hostingName}`)
-    const hosting = new Hosting(hostingName, socket)
+    const hosting = new Hosting(hostingName)
     return hosting.loadRemote()
-  }
-
-  static listLocal (socket) {
-    return socket.settings.listHosting()
   }
 
   static listFromProject () {
@@ -186,14 +203,11 @@ class Hosting {
   }
 
   // list all hostings (mix of locally definde and installed on server)
-  static async list (socket) {
+  static async list () {
     debug('list()')
-    if (!socket) {
-      const projectHostings = Hosting.listFromProject()
-      debug('projectHostings', projectHostings)
-      const promises = projectHostings.map((hosting) => Hosting.get(hosting.name))
-      return Promise.all(promises)
-    }
+    const projectHostings = Hosting.listFromProject()
+    debug('projectHostings', projectHostings)
+    return Promise.all(projectHostings.map((hosting) => Hosting.get(hosting.name)))
   }
 
   static getDirectories () {
@@ -222,10 +236,12 @@ class Hosting {
     debug('setRemoteState', hosting.name)
     if (hosting && typeof hosting === 'object') {
       this.existRemotely = true
-      this.name = hosting.name
-      this.description = hosting.description
-      this.domains = hosting.domains
-      this.config.browser_router = hosting.config.browser_router || false
+      this.remote.name = hosting.name
+      this.remote.description = hosting.description
+      this.remote.domains = hosting.domains
+      this.remote.config = hosting.config || {}
+      this.remote.config.browser_router = hosting.config.browser_router || false
+      this.remote.auth = hosting.auth
       this.isUpToDate = await this.areFilesUpToDate()
     } else {
       this.existRemotely = false
@@ -247,28 +263,18 @@ class Hosting {
 
   loadLocal () {
     debug('loadLocal()')
-    let localHostingSettings = {}
-    if (this.socket) {
-      if (this.socket.settings.loaded) {
-        localHostingSettings = this.socket.settings.getHosting(this.name)
-      }
-    } else {
-      localHostingSettings = session.settings.project.getHosting(this.name)
-    }
+    const localHostingSettings = session.settings.project.getHosting(this.name)
 
-    if (!localHostingSettings) {
-      return
-    }
+    if (localHostingSettings) {
+      if (Object.keys(localHostingSettings).length > 0) {
+        this.existLocally = true
+        this.src = localHostingSettings.src
+        this.cname = localHostingSettings.cname
+        this.auth = localHostingSettings.auth
+        this.path = path.join(session.projectPath, this.src, path.sep)
+        this.url = this.getURL(this.name)
 
-    if (Object.keys(localHostingSettings).length > 0) {
-      this.existLocally = true
-      this.src = localHostingSettings.src
-      this.cname = localHostingSettings.cname
-      this.path = path.join(session.projectPath, this.src, path.sep)
-      this.url = this.getURL(this.name)
-
-      if (localHostingSettings.config && localHostingSettings.config.browser_router) {
-        this.config.browser_router = localHostingSettings.config.browser_router
+        this.config = localHostingSettings.config || {}
       }
     }
   }
@@ -467,7 +473,7 @@ class Hosting {
   }
 
   getCNAME () {
-    return _.find(this.domains, (domain) => domain.indexOf('.') !== -1)
+    return _.find(this.remote.domains, (domain) => domain.indexOf('.') !== -1)
   }
 
   getCnameURL () {