+++ packages/lib-js-client/jest.config.js
@@ -1,0 +1,14 @@
+module.exports = {
+  "collectCoverage": true,
+  "collectCoverageFrom": [
+    "src/**/*.{ts}",
+  ],
+  "transform": {
+    ".*": `${__dirname}/../../node_modules/ts-jest/preprocessor.js`
+  },
+  "testRegex": "(\\.(spec))\\.(ts)$",
+  "moduleFileExtensions": [
+    "ts",
+    "js"
+  ]
+}